+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added [hdbpp-healthcheck device server]
+
+- Check on more than 1 health endpoints.
+- Configure the endpoints through a device server property.
+
 ## [0.2.0] - 2020-03-06
 
 ### Added [general]